generics semantic analysis half-finished (desc)
Okay, I redid the generics (again!)

it almost works perfectly, e.g. the following code is immediately marked as incorrect:
```
GenType<int> a;
GenType<GenType<GenType<int>>> b;
a = b;
```

What's left to make is to store the given generic arg to the param in the type itself. A property with type T on a generic type doesn't really know what its actual type is at the moment, and that's a problem!